requirements.txt
asgiref==3.4.1
beautifultable==0.8.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fandogh-cli==1.26.7
idna==3.2
kavenegar==1.1.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyotp==2.6.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
websocket-client==1.2.1
